Health survey data from Missouri residents (BRFSS).

The full dataset containing basic survey information, demographics, health information, tobacco usage, and cognitive decline prevalence from the years 2023 and 2024.

Usage

missouri_brfss

Format

A data frame with 14,528 rows and 34 variables:

SEQNO

key of the dataset

IDAY

interview day

IMONTH

interview month

IYEAR

interview year

MARITAL

marital status

EDUCA

education level

VETERAN3

veteran status

EMPLOY1

employment status

CHILDREN

number of children in household

INCOME3

income level

PREGNANT

pregnancy status

WEIGHT2

reported weight in pounds

CRACE1

child non-hispanic race including multiracial

ASTHMS1

asthma status

SMOKER3

general smoking status

CAGEG

child age

AGE80

age

CHCOCNC1

melanoma or other types of cancer

ALCDAY4

days in the past 30 had alcohol

MARJSMOK

marijuana usage

GENHLTH

general health

MENTHLTH

mental health

SMOKE100

smoked at least 100 cigarettes

SMOKE100_IMP1

1st imputation of SMOKE100

SMOKE100_IMP2

2nd imputation of SMOKE100

SMOKE100_IMP3

3rd imputation of SMOKE100

ECIGNOW

e-cigarettes usage

ECIGNOW_IMP1

1st imputation of ECIGNOW

ECIGNOW_IMP2

2nd imputation of ECIGNOW

ECIGNOW_IMP3

3rd imputation of ECIGNOW

CIMEMLO1

difficulties with thinking or memory

CIMEMLO1_IMP1

1st imputation of CIMEMLO1

CIMEMLO1_IMP2

2nd imputation of CIMEMLO1

CIMEMLO1_IMP3

3rd imputation of CIMEMLO1

Source

https://www.cdc.gov/brfss/index.html